Refact and unittests
Refact the code and add unit tests, before #7 issue modifications
New line char fix

New line char fix

New line char fix

New line char fix

New line char fix
