+++ tools/__init__.py
@@ -1,2 +1,4 @@
 from .shellhelper import get_pipe_output, get_external_execution_time
 from .gitstatistics import GitStatistics
+from .datacollector import GitDataCollector
+from .configuration import Configuration, ConfigurationException, UsageException
